requirements.txt
absl-py==0.12.0
ai-benchmark==0.1.2
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.9.3
boto3==1.17.9
botocore==1.20.19
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
chart-studio==1.1.0
configparser==5.0.1
cssselect==1.1.0
cycler==0.10.0
dash==1.19.0
dash-auth==1.4.1
dash-bootstrap-components==0.11.3
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
dash-treeview-antd==0.0.1
dash-uploader==0.4.1
dateparser==1.0.0
demjson==2.2.4
docker-pycreds==0.4.0
fake-useragent==0.1.11
Flask==1.1.2
Flask-Compress==1.9.0
Flask-SeaSurf==0.3.0
flatbuffers==1.12
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.13
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.2
numpy==1.20.3
opt-einsum==3.3.0
parse==1.19.0
pathtools==0.1.2
Pillow==8.2.0
plotly==4.14.3
promise==2.3
protobuf==3.15.2
psutil==5.8.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.1.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
retrying==1.3.3
rq==1.7.0
rsa==4.7.2
s3transfer==0.3.4
sentry-sdk==0.20.3
shortuuid==1.0.1
six==1.16.0
smmap==3.0.5
soupsieve==2.2.1
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.0
typing-extensions==3.10.0.0
tzlocal==2.1
ua-parser==0.10.0
urllib3==1.26.4
w3lib==1.22.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
